+++ src/lib/server/db/schema/entry.ts
@@ -30,7 +30,8 @@ export const entries = pgTable(
 		license: licenseEnum('license').notNull(),
 		// Files
 		mediaId: uuid('media_id').notNull(),
-		thumbnailId: uuid('thumbnail_id').notNull(),
+		thumbnailId: uuid('thumbnail_id'), // Stale for now
+		previewId: uuid('preview_id').notNull(),
 		proofId: uuid('proof_id'),
 		...timestamps
 	},
@@ -67,6 +68,10 @@ export const entriesRelations = relations(entries, ({ one, many }) => ({
 	}),
 	thumbnail: one(files, {
 		fields: [entries.thumbnailId],
+		references: [files.id]
+	}),
+	preview: one(files, {
+		fields: [entries.previewId],
 		references: [files.id]
 	}),
 	proof: one(files, {