+++ src/lib/types/AzureStorageContainer.ts
@@ -3,6 +3,8 @@ export const azureStorageContainer = [
 	'video-entries',
 	'image-entries',
 	'audio-entries',
-	'avatar'
+	'avatars',
+	'thumbnails',
+	'previews'
 ] as const;
 export type AzureStorageContainer = (typeof azureStorageContainer)[number];