+++ src/routes/(app)/entries/create/_components/FinishSection.svelte
@@ -7,6 +7,8 @@
 	import { createEntrySchema, type CreateEntrySchema } from '$lib/schemas/entry';
 	import { getContext } from 'svelte';
 	import type { SuperForm } from 'sveltekit-superforms';
+	import SubmittingTo from './SubmittingTo.svelte';
+	import Divider from '$lib/components/Divider.svelte';
 	let { isValid = $bindable() }: { isValid: boolean } = $props();
 	const licenses: StyledSelectItem[] = licenseCode.map((code) => ({
 		label: code.toUpperCase(),
@@ -21,6 +23,8 @@
 </script>
 
 <section>
+	<SubmittingTo />
+	<Divider class="my-xl" />
 	<div class="flex flex-col gap-xs">
 		<TextField field="title" label="Title" labelProps={{ class: 'text-2xl font-bold' }} />
 		<p class="text-shade-300">What do you call your masterpiece?</p>